% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_function.R
\name{bc_prs}
\alias{bc_prs}
\title{This function is to corect the bias of cross-trait PRS according to Theorem 1 and 4}
\usage{
bc_prs(raw_estimator, n_train, p_indep, h2_trait1, h2_trait2, n_overlap = 0,
  n_test)
}
\arguments{
\item{raw_estimator}{the raw genetic correlation estimator in PRS}

\item{n_train}{the sample size of training GWAS}

\item{p_indep}{the number of independent genetic variants}

\item{h2_trait1}{heritability of the trait in training GWAS}

\item{h2_trait2}{heritability of the trait in tetsing GWAS}

\item{n_overlap}{the number of overlaped samples between training and testing GWAS, default=0}

\item{n_test}{the sample size of testing GWAS (only useful when n_overlap >0,default=0)}
}
\value{
bias_corrected_estimator: raw bias-corrected genetic correlation estimator
}
\description{
This function is to corect the bias of cross-trait PRS according to Theorem 1 and 4
}
\examples{

##toy example 
bc_prs(raw_estimator=0.1,n_train=50000,p_indep=500000,
h2_trait1=0.5,h2_trait2=0.5,n_overlap=0,n_test=0)
}
\keyword{bc_prs}
