% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_function.R
\name{bc_prs2}
\alias{bc_prs2}
\title{This function is to corect the bias of cross-trait PRS according to Theorem 2,3 and Proposition S4}
\usage{
bc_prs2(raw_estimator, n_train1, n_train2, p_indep, h2_trait1, h2_trait2,
  n_overlap = 0)
}
\arguments{
\item{raw_estimator}{the raw genetic correlation estimator in PRS}

\item{n_train1}{the sample size of training GWAS1}

\item{n_train2}{the sample size of training GWAS2}

\item{p_indep}{the number of independent genetic variants}

\item{h2_trait1}{heritability of the trait in training GWAS1}

\item{h2_trait2}{heritability of the trait in training GWAS2}

\item{n_overlap}{the number of overlaped samples between the two training GWAS, default=0}
}
\value{
bias_corrected_estimator: raw bias-corrected genetic correlation estimator
}
\description{
This function is to corect the bias of cross-trait PRS according to Theorem 2,3 and Proposition S4
}
\examples{

##toy example 
bc_prs2(raw_estimator=0.1,n_train1=100000,
n_train2=100000,p_indep=300000,h2_trait1=0.5,h2_trait2=0.5,n_overlap=0)
}
\keyword{bc_prs2}
